Add iterator-visitor classes (and test) to allow walking and modifying gcode
The filter classes have two drawbacks:

* The filters build two kinds of state as they pass over gcode, but they only need one:
  . the work they care about (eg, collecting a group of linear segments and converting them to an arc)
  . their position in the gcode in support of the work they care about. This state can be tracked elsewhere and not repeated in each filter.
* The filters do not differentiate between the two flavors of GCoder layers: with or without extrusion
  . When GCoder encounters pure movement commands, usually in the prologue and epilogue stanzas from the slicer, it groups them as a "layer", but this group is not represented in the object's all_zs property.
  . Thus, for GCoder objects, len(self.all_layers) != len(self.all_zs)
  . Operations on gcode that require an awareness of the *printed* layer need to track additional state, but this is not the responsibility of a filter.

To address these, the iterator class tracks the visitor's place in the gcode and then passes that information to the visitor at each step so that the visitor only needs to keep state that directly helps its task. This improves the "signal to boilerplate" ratio. Part of the state passed to the visitor is a bool that represents whether or not the layer 'is_printed'.